requirements.txt
opencv-python
numpy
pandas
scipy
scikit-image
matplotlib
pyarrow
pyyaml
tqdm
soundfile
